+++ licovani/multiPOC_Ai1.h
@@ -4,7 +4,7 @@
 #include <QCheckBox>
 int kompletni_slicovani(cv::VideoCapture& cap,
                         cv::Mat& referencni_snimek,
-                        int cislo_posunuty,
+                        double cislo_posunuty,
                         int& iterace,
                         double& oblastMaxima,
                         double& uhel,
